Updated docker compose, small fix and new Extra Costs feature
I have a suspicion something is changing with the key server and a validation after 15 minutes isn't good.  Let's not validate on calls to /me.  I've been running off this for several hours without losing my session.

Add viewing, adding and editing of ExtraCosts.

Add viewing and searching of Items